Add edsl module
